requirements.txt
dnspython3
ipaddress
